Move reporting infrastructure into core
