+++ pikelet-cli/Cargo.toml
@@ -20,7 +20,6 @@ gui = ["pikelet-gui"]
 
 [dependencies]
 annotate-snippets = "0.6.1"
-itertools = "0.8"
 pikelet = { path = "../pikelet" }
 pikelet-gui = { path = "../pikelet-gui", optional = true }
 pretty = "0.9"
