+++ pikelet-cli/src/lib.rs
@@ -1,7 +1,6 @@
 use std::error::Error;
 
 pub mod repl;
-pub mod reporting;
 
 /// The Pikelet command line interface.
 #[derive(structopt::StructOpt)]
