+++ pikelet/Cargo.toml
@@ -15,7 +15,9 @@ license = "Apache-2.0"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+annotate-snippets = "0.6.1"
 im = "14.2"
+itertools = "0.8"
 lalrpop-util = "0.17"
 pretty = "0.9"
 regex = "1.3"
